pyproject.toml
[tool.poetry]
name = "creaturemmo-bot"
version = "0.1.0"
description = "Simple CLI Tool for grinding creatures"
authors = ["js-bruno <brunocebrsilva@gmail.com>"]
packages = [
    { include = "creature_core", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.7.0"
PyAutoGUI = "^0.9.53"
Pillow = "^9.5.0"
opencv-python = "^4.7.0"
pytesseract = "^0.3.10"
pytest = "^7.3.0"
pytest-bdd = "^6.1.1"
pytest-mock = "^3.10.0"
python-xlib = "^0.33"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
